+++ CHANGELOG.md
@@ -6,12 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased] - 2024-01-18
-### Removed
-- Removed Grid item from `Layout`.
 
 ### Changed
 - Increased width of drawer in `NavMenu` to 60%.
 - Increased font size of `ListItemButton` to 1.2em.
+- Set padding options in sigle pages
 
 ## [v15.14.0] - 2024-01-16